+++ ConsoleApp3/Objet/Electro-menager.cs
@@ -8,7 +8,7 @@ namespace ConsoleApp3
 {
     class Electro_menager : Objet
     {
-        public Electro_menager(string objetid, string Type, string description, string lieuStock, int montant, string état) : base(objetid, Type, description, état, lieuStock, montant)
+        public Electro_menager(string objetid, string Type, string description, string lieuStock, double montant, string état) : base(objetid, Type, description, état, lieuStock, montant)
         {
             this.Type = "Cuisine";
         }
