+++ ConsoleApp3/Objet/Lave-Linge.cs
@@ -8,7 +8,7 @@ namespace ConsoleApp3
 {
     class Lave_Linge:Electro_menager
     {
-        public Lave_Linge(string objetid, string Type, string description, string lieuStock, int montant, string etat) : base(objetid, Type, description, lieuStock, montant, etat)
+        public Lave_Linge(string objetid, string Type, string description, string lieuStock, double montant, string etat) : base(objetid, Type, description, lieuStock, montant, etat)
         {
             this.Description = "Lave_Linge";
         }
