+++ ConsoleApp3/Objet/Matelas.cs
@@ -8,7 +8,7 @@ namespace ConsoleApp3
 {
     class Matelas : MobilierChambre
     {
-        public Matelas(string objetid, string Type, string description, string lieuStock, int montant, string etat) : base(objetid, Type, description, lieuStock, montant, etat)
+        public Matelas(string objetid, string Type, string description, string lieuStock, double montant, string etat) : base(objetid, Type, description, lieuStock, montant, etat)
         {
             this.Description = "Matelas";
         }
