+++ ConsoleApp3/Objet/MobilierChambre.cs
@@ -9,7 +9,7 @@ namespace ConsoleApp3
     class MobilierChambre : Objet
     {
         
-        public MobilierChambre(string objetid, string Type, string description, string lieuStock, int montant, string etat) : base(objetid, Type, description,etat, lieuStock, montant)
+        public MobilierChambre(string objetid, string Type, string description, string lieuStock, double montant, string etat) : base(objetid, Type, description,etat, lieuStock, montant)
         {
             this.Type = "Chambre";
         }
