+++ ConsoleApp3/Objet/Réfrigérateur.cs
@@ -8,7 +8,7 @@ namespace ConsoleApp3
 {
     class Réfrigérateur : Electro_menager
     {
-        public Réfrigérateur(string objetid, string Type, string description, string lieuStock, int montant, string etat) : base(objetid, Type, description, lieuStock, montant, etat)
+        public Réfrigérateur(string objetid, string Type, string description, string lieuStock, double montant, string etat) : base(objetid, Type, description, lieuStock, montant, etat)
         {
             this.Description = "Réfrigérateur";
         }
